+++ src/test/resources/andrew-cart.json
@@ -1,0 +1,1 @@
+{"cartName":"andrew-cart","realItems":[{"weight":1560.0,"name":"Audi","price":32026.9}],"virtualItems":[{"sizeOnDisk":20000.0,"name":"Windows","price":11.0}],"total":38445.479999999996}
